cc: fix int128 bugs — ternary, stack-spilled args, global loads + tests
Fix four int128 codegen bugs and add comprehensive test coverage:

Bug 1: Ternary ?: with int128 operands used Select instruction which
caps at 64 bits. Add size<=64 guard to force branch+phi path.

Bug 2: x86_64 stack-spilled int128 call args pushed only 8 bytes
instead of 16. Add int128 branch in push_stack_args, fix stack qword
counting in classify_call_args, and copy incoming stack int128 params
to local slots in the callee prologue (store_args_to_stack).

Bug 3: aarch64 stack-spilled int128 call args allocated 8 bytes
instead of 16. Fix allocation calculation and add LDP+STR pair for
16-byte stack arg stores.

Bug 4: Global int128 loads on x86_64 double-dereferenced the GOT
address (treating the int128 value as a pointer). Add Loc::Global
branch in emit_load that loads the address without dereferencing.

Add 11 new tests: ternary, many_args (stack spill), divmod, globals,
ptr_deref, compound_assign, shift_boundaries, float_convert,
struct_array, inc_dec, optimized_mega.

+++ cc/arch/aarch64/features.rs
@@ -389,7 +389,7 @@ impl Aarch64CodeGen {
             Loc::Imm(v) => {
                 self.push_lir(Aarch64Inst::Mov {
                     size: mov_size,
-                    src: GpOperand::Imm(*v),
+                    src: GpOperand::Imm(*v as i64),
                     dst: scratch,
                 });
             }
@@ -483,7 +483,7 @@ impl Aarch64CodeGen {
             Loc::Imm(v) => {
                 self.push_lir(Aarch64Inst::Mov {
                     size: src_size,
-                    src: GpOperand::Imm(v),
+                    src: GpOperand::Imm(v as i64),
                     dst: scratch,
                 });
             }
@@ -567,7 +567,7 @@ impl Aarch64CodeGen {
             }
             Loc::Imm(v) => {
                 // Use emit_mov_imm which handles large immediates with movz/movk
-                self.emit_mov_imm(scratch, v, src_size.bits());
+                self.emit_mov_imm(scratch, v as i64, src_size.bits());
             }
             _ => return,
         }
@@ -644,7 +644,7 @@ impl Aarch64CodeGen {
             }
             Loc::Imm(v) => {
                 // Use emit_mov_imm which handles large immediates with movz/movk
-                self.emit_mov_imm(scratch, v, src_size.bits());
+                self.emit_mov_imm(scratch, v as i64, src_size.bits());
             }
             _ => return,
         }
